refactor: 重整專案架構
將程式移到 `booflow/__init__.py` 中，這樣比較簡潔
